+++ treeview.ts
@@ -94,7 +94,7 @@ export async function showTree() {
     asset.readAsset(PLUG_NAME, "assets/icons/x-circle.svg"),
   ]);
 
-  const { currentPage, nodes } = await getPageTree(config.pageExcludeRegex);
+  const { currentPage, nodes } = await getPageTree(config);
   const customStyles = await getCustomStyles();
 
   const treeViewConfig = {